Add the ability to run a profile for a given duration instead of a number of iterations.
The duration is passed in human readable format (example : "1d 5h 10m") and will be converted to minutes.
If the duration is set to 0, the number of iterations will be used instead.

Use thread specific semaphores to handle concurrency and exit stress runs correctly.
Remove unnecessary stuff in StressContext that were already available in mainArguments.

It was impossible for multiple threads to exit because of the shared semaphore and because they were all trying to get all the permits before exiting.
Then, with a shared semaphore, work distribution could be uneven due to some threads acquiring more permits than others.
Using a local semaphore will ease picking the right settings between concurrency and number of threads, along with allowing a clean exit for multi threaded runs.

Add documentation for running stress tests for a specific duration